random 和 作用域（scope）

+++ C++大学教程练习题/第06章 函数/cpp_11_random.cpp
@@ -2,12 +2,15 @@
 #include <random>
 #include <ctime>
 using namespace std;
+
 int main()
 {
-    default_random_engine engine(time(0));
+    default_random_engine engine(static_cast<unsigned int>(time(0)));
+
     uniform_int_distribution<long> randomInt(-101, 6); // <>表名其为类模板
     cout << randomInt(engine) << endl;
-    uniform_real_distribution<float> randomReal(-10000, 100);
+
+    uniform_real_distribution<double> randomReal(-10000, 100);
     cout << randomReal(engine) << endl;
     return 0;
 }
